Move ComposableModel class hierarchy into its own file. Add unit test for Composable model. Remove some dead code. Change: 126918604
